+++ src/map.rs
@@ -685,6 +685,15 @@ where
     }
 }
 
+impl<K, V> Extend<(K, V)> for NEMap<K, V>
+where
+    K: Eq + Hash,
+{
+    fn extend<I: IntoIterator<Item = (K, V)>>(&mut self, iter: I) {
+        self.inner.extend(iter);
+    }
+}
+
 #[cfg(test)]
 mod test {
     use std::num::NonZeroUsize;